Rename libresource0-*.pc to libresource-*.pc
In general guideline is *.so -> *.pc.

Let's clarify; when debian mentiones libcanberra0, it means the SONAME
of the library is libcanberra.so.0. When libcanberra1 is released,
libcanberra.so remains unmodified, so when packages link to it, the
linker will pick libcanberra.so.1, that's why libcanberra-dev doesn't
contain the '0' in it, and neither does libcanberra.pc.

Contast this with libdbus-1-3; the soname would be libdbus-1.so.3. So
when libdbus-1-4 is released, only recompilation is needed (ABI
compatiblity broken), when libdbus-2-0 is released, the source code
might require modifications (API compatibility broken).

Therefore, the logical names of the pc files are dbus-1.pc and
libcanberra.pc, following dbus-1.so and libcanberra.so.

If libcanberra decides to break API, should be renamed to libcanberra-1.

Let's do the same with libresource.

Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>